Home  controller ProductModel
Wrote a controller that stands between the Models and the home view. With this fix, I changed the architecture of the app a little bit. Each of the data models is present in the Models folder.  The controller is responsible to handle all of the interactions between view and model.

I also added some html in the view file (index.php). Note that, all of the logic in the view is being written in top level of the file between the <?php //..// ?>