Removed framework stuff + re-factored DatabaseClient
Removed all of the forbidden folders and files from within the "Vendor" folder. Removed "symphony" and "React" together with all of the other "frameworks"

Re-factored the DatabaseClient class and it's methods.  Instead of JSON and HTTP requests on the different api endpoints of supabase, now the communication to database is based on pgSQL.
The methods remain the same, however re-factored. Now raw SQL statements(queries) are used to query the database.