Shopping cart, Shop About models new methods for DatabaseClient
Added a new method, that updates a table based on table name and column name and item id.

Also added a small model, that handles the shop about info

Added a model that represents the Shopping cart. This model is supposed to handle when user is adding or removing products from the shopping cart. The model is to be finished in the future.