Shopping cart back-end logic
Wrote the ShoppingCart model, that handles all of the shopping cart methods, such as add to cart, remove from cart, get all items.
Wrote the ShoppingCartController which is in charge of handling all of the HTTP request from and to the shopping cart back-end.
Wrote the AJAX methods for dynamically interacting with shopping cart from the front end. The AJAX is currently in the index.php, but is to be moved into a separate file.