Dynamic routes ProductOverviewController
I re-factored the code logic so it's prepared for the dynamic routing of the different products. Changed the path to the product.php screen in the routes.php file. The dynamic routing is based on url parameter. When clicking on a picture, the url get's is id as a parameter. In the produc.php class the url parameter is retreived with the $_GET['id'] method and passed to the ProductOverviewController getProduct() method.
The ProductOverviewController class is in charge of maintaing the in-product view (product.php). Currently the class has only one method which is getProduct() and the method returns the value from the DatabaseClient class, getSpecificId_FromTable();

I also changed the product.php view to display the product's inormation on screen.

Added a new Model "CartProduct.php". It will be in charge of handling the fully composed product (photo, frame, size etc.).
Changed the name of the PhotoModel class to just Photo.