Email sender class+ajax
Wrote the email sender class and the EmailSenderController. The EmailSender class is a model that holds the method for sending email. The EmailSenderController is in charge of getting the POST request sent async from the client side and call the sendEmail method from the EmailSender model.
Also wrote the javascript requests for contacting the EmailSenderController. The client side uses a custom route for contacting the EmailSenderController.