Shopping cart back-end
Most of the shopping cart back-end is done. Now the user can easily add the photo object to the cart (session) and when they do that, they automatically get transffered back to home page.
What remains to be done is implementing the front end of the shopping cart, together with the quantity control system. Then we can try and implement the whole system with fram, size and so on.