AdminLoginController, admin login view
Added a a new method in the DatabaseClient AuthenticationLogIn(), that is supposed to handle the sign-in of an admin. The method takes a string as anrgument, which in this context will be the hashed user password checks if the password exists in the database.
I also wrote a AdminLoginController that handles the interaction from the admin-login view. The controller currently has one method signIn() which takes a string as an argument, which is supposed to be the user's password. The method is hashing the password and sends it to the AuthenticationLogIn method(), part of the DatabaseClient method.
Added a new route for the admin-login page.