AdminPannel controller and shopping cart update
Wrote the AdminPannelController, also added four methods:
getShopAbout() method gets the shop about text, updateShopAboutInfo() which updates the shop about text and updateShopOwnerTel() method which updates the owner telephone number, also a method for getting all orders getAllOrders(). In the controller there is an extra method called handlePannelInteractions(), which is in charge of handling the async updates coming from the client side. Currently the client side communicates to the controller using AJAX, the methods for that can be found in the adminPannel.js file present in the Public folder.
Updated the view for admin-panel.php, added some simple input fields for handling the data inputs for the different updates. Also added a table that displays all of the different orders.

Fixed the front-end of the shopping cart, now the shopping displays the items added to it in a nicely presented way.