Fixed authentication
Fixed the authentication and discovered the problems were caused because the argument for password and username were passed in reverse to the method frim the AdminLoginController.
Added a thin layer of security, making sure the admin pannel is not reachable if the user doens't log in. Did that using the $_SESSION[authState]. The router is checking if the state is set and if's not, the route normally leading to the admin pannel leads to 404 page.