Added a quantity tracking system to shopping cart
Modified the CartProduct class to include quantity property. also added the id to the Photo class. This way the shopping cart can track if an item with the same id exists and if it does, it increments it's quantity.
Modified the addtoCart method part of the shoppingCart class. The new version of the method checks if a product with the same id as the product being added exists and if it does it increments the quantity.
Changed how the items are displayed in the shopping cart presentation.